// src/models/dtos/customer.model.ts
export enum Customer {
    TIM = "TIM",
    TEAMSYSTEM = "TeamSystem",
    SOGEITSOLUTIONS = "Sogeit Solutions",
    FOODLOVERY = "FoodLovery",
    ENGINEERING = "Engineering",
    LIVEBOX = "Livebox",
    DEEP_CONSULTING = "Deep Consulting S.r.l.",
    GSE = "GSE - Gestore Servizi Energetici S.p.A.",
    BIP = "BIP - Business Innovation Partners",
    INTERNO = "Interno"
}